requirements.txt
asgiref==3.7.2
awsebcli==3.20.3
blessed==1.20.0
botocore==1.35.99
cement==2.10.14
certifi==2024.8.30
charset-normalizer==2.1.1
colorama==0.4.6
Django==3.2.5
django-cors-headers==3.14.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.7
jmespath==1.0.1
joblib==1.4.2
numpy==1.21.6
packaging==24.2
pathspec==0.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.3
semantic-version==2.10.0
setuptools==65.7.0
six==1.17.0
sqlparse==0.4.4
termcolor==2.5.0
threadpoolctl==3.6.0
urllib3==1.26.18
wcwidth==0.2.13
whitenoise==6.2.0
